Make sure of separator
We use the separator later on to determine where the yaml document starts.
`eyaml edit` adds the separator to new (non-existing) files by itself but
since we want to create the file before in order to diff later the separator
needs to be added in order to get a valid document.